# combined read/write methods for an improvement in performance