--- a/Adafruit_SPIDevice.h
+++ b/Adafruit_SPIDevice.h
@@ -88,6 +88,8 @@ public:
   void transfer(uint8_t *buffer, size_t len);
   void beginTransaction(void);
   void endTransaction(void);
+  void beginTransactionWithAssertingCS(void);
+  void endTransactionWithDeassertingCS(void);
 
 private:
   SPIClass *_spi;