@eringerli 2022-04-17
I reorganised my repo, so this pull-request ist a copy of #87