@ladyada 2022-04-17
ok lets split this again into a PR for begin/end transaction changes and then the buffering changes!