@eringerli 2022-04-18
A huge gain in reading performace is already in #90: instead of reading each byte individually in a loop, do it in one go. This PR adds the combination of reading/writing to it, which removes the addition round trip of transation management in between.