@eringerli 2022-04-21
This is because the history looks like this: ![Screenshot_20220421_062914](https://user-images.githubusercontent.com/48175951/164372303-18c862a4-9d8c-42d0-83f8-0f4139ec93e3.png)

PR #90 is for the branch `streamlined-transaction-management`, PR #88 is for the branch `read-write-combine`

It is too much work for me too back port the caching stuff without the CS/transaction stuff, as this just makes no sense.