@eringerli 2022-04-21
The buffering stuff is this:

```
    size_t lenBuffer = prefix_len + len;
    if (lenBuffer < minBufferSizeToMalloc) {
      uint8_t tmpBuffer[lenBuffer];

      memcpy(tmpBuffer, prefix_buffer, len);
      memcpy(tmpBuffer + prefix_len, buffer, len);

      write_and_read(tmpBuffer, lenBuffer);
    } else {
      auto tmpBuffer = new uint8_t[lenBuffer];

      memcpy(tmpBuffer, prefix_buffer, len);
      memcpy(tmpBuffer + prefix_len, buffer, len);

      write_and_read(tmpBuffer, lenBuffer);

      delete buffer;
    }
```

The CS/transaction is this:
```
  beginTransactionWithAssertingCS();
  transfer(buffer, len);
  endTransactionWithDeassertingCS();
```

So no, #90 doesn't contain any buffering changes.