@ladyada 2022-05-02
id like the two conceptual changes: buffering the tx's  /  changing the CS pin and beginTransaction, to be seperated into two PRs so they can be reviewed and merged sepearately in case we need to revert changes!