@ladyada 2022-05-02
not sure whats up with the two PRs but we see changes to spi data/transactions in 
https://github.com/adafruit/Adafruit_BusIO/pull/90/files

![image](https://user-images.githubusercontent.com/1214161/166185535-ab082116-1239-4155-a211-c677a17dd4b8.png)

maybe check the branches?