@eringerli 2022-05-02
Yes, #90 is about transactions, #88 is about data changes. The numbers are  bit confusing, as #88 is based on #90, but that was just lazy me that forcepushed and rewrote a PR. I split the transaction stuff out and made a new PR, which got number 90