@ladyada 2022-05-02
right, please start with a pr that is *just changing CS and begintransaction in all locations* and we'll go from there