@ladyada 2022-05-02
correct, and refactoring is super fun and often introduces bugs because so much has changed!