@eringerli 2022-05-02
Should I split the other PR further into logical commits? Would that be a sensible compromise?